backend/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
aiomysql==0.2.0
cryptography==42.0.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.0
google-cloud-logging==3.9.0
google-auth==2.27.0
prometheus-client==0.19.0
structlog==24.1.0
tenacity==8.2.3
sentry-sdk[fastapi]==1.40.0
python-dotenv==1.0.0
